+++ packages/core/src/canvas/flatten.ts
@@ -22,15 +22,14 @@ function nodesToVectorProps(
   nodes: SceneNode[],
   makeNodePath: NodePathFactory
 ): VectorFlattenProps | null {
-  const path = new renderer.ck.Path()
+  const path = new renderer.ck.PathBuilder()
   for (const node of nodes) {
     const nodePath = makeNodePath(renderer, graph, node)
     if (!nodePath) {
       path.delete()
       return null
     }
-    nodePath.transform(nodePathTransform(renderer, node))
-    path.addPath(nodePath)
+    path.addPath(nodePath, nodePathTransform(renderer, node))
     nodePath.delete()
   }
 
@@ -41,8 +40,9 @@ function nodesToVectorProps(
   }
 
   path.transform(renderer.ck.Matrix.translated(-bounds[0], -bounds[1]))
-  const vectorNetwork = parseSVGPath(path.toSVGString())
-  path.delete()
+  const immutablePath = path.detachAndDelete()
+  const vectorNetwork = parseSVGPath(immutablePath.toSVGString())
+  immutablePath.delete()
 
   return {
     name: 'Flatten',